monitoring/requirements.txt
# Monitoring System Dependencies

# Dashboard
dash==2.18.1
dash-bootstrap-components==1.6.0
plotly==5.24.1
pandas==2.2.3

# CLI
rich==13.9.4
typer==0.15.1
questionary==2.0.1

# Metrics & Storage
sqlite-utils==3.37
python-dotenv==1.0.1

# GitHub Integration
PyGithub==2.5.0

# Utilities
requests==2.32.3
psutil==6.1.0
gitpython==3.1.43
